Enemies Creation - Part 1
This is where I finish for today, most likely. I added 3 rows of 7 enemies (ships). 

I had an idea that we could possibly turn this into a minigame for Stellarios, where you need to "destroy drone ships" to defend the galactic capital, and there would be different levels. 

Check out `Readme.md` for the thread