#4
#4
All of the enemy game objects in the prefab had multiple copies of the enemy.cs script, so I'm going through the tutorials again that are relevant to this issue to rewrite the code (this also serves the purpose of helping reinforce what we're doing) and, in conjunction with: https://github.com/gizmotronn/unity-intro/issues/1 & https://github.com/irisdroidology/python-learning/issues/3, I'm using the Unity intro tutorials to also further enhance my knowledge of the Unity editor and the project (similar to what I'm doing with the Python-Learning issue linked). I also got a message from Zenva at liam@acordsoftware.tech  regarding this. It may have just been a problem with the code and the game objects being incorrectly configured.