https://github.com/Gizmotronn/space-attackers/issues/4
Enemies can now shoot back and are destroyed upon contact with a player missile! This was thanks to properly adding the `enemy.cs` script and its attributes to the enemy prefab gamObjects in Unity! See the attached image in the commit title for more ^^

Got a problem where the player isn't being destroyed upon contact with the enemy missile but we will fix that later